+++ prept/commands/new.py
@@ -127,7 +127,7 @@ def new(
             output=output,
             variables=variables,
         )
-        tp = boilerplate.template_provider(**boilerplate.template_provider_params) if boilerplate.template_provider else None
+        tp = boilerplate._get_template_provider()
         engine = boilerplate.engine
 
         if engine: